services/asset-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.26.1
structlog==23.2.0
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
aiofiles==23.2.1
